messed around trying to create an actual level
added some textures. messed around with a level...the block doesn't fall immediately. need to work on that a bit more I guess.
Think I need to adjust the positions it ends up in to round to nearest half probably so it will fall in that hole in the middle...or maybe the sizes need adjusted? not sure.
probably good enough for a prototype